+++ server_extensions_extra/src/main/java/com/generalbytes/batm/server/extensions/extra/dash/sources/coinmarketcap/CMCMetaData.java
@@ -1,0 +1,4 @@
+public class CMCMetaData {
+    public long timestamp;
+    public String error;
+}
