+++ server_extensions_extra/src/main/java/com/generalbytes/batm/server/extensions/extra/dash/sources/coinmarketcap/CMCQuote.java
@@ -1,0 +1,9 @@
+import java.math.BigDecimal;
+public class CMCQuote {
+    public BigDecimal price;
+    public BigDecimal volume_24h;
+    public BigDecimal market_cap;
+    public BigDecimal percent_change_1h;
+    public BigDecimal percent_change_24h;
+    public BigDecimal percent_change_7d;
+}
