+++ server_extensions_extra/src/main/java/com/generalbytes/batm/server/extensions/extra/dash/sources/coinmarketcap/CMCQuotes.java
@@ -1,0 +1,4 @@
+import com.generalbytes.batm.server.extensions.extra.dash.sources.coinmarketcap.CMCQuote;
+public class CMCQuotes {
+    public CMCQuote USD;
+}
