+++ server_extensions_extra/src/main/java/com/generalbytes/batm/server/extensions/extra/dash/sources/coinmarketcap/CMCTicker.java
@@ -1,97 +1,69 @@
 package com.generalbytes.batm.server.extensions.extra.dash.sources.coinmarketcap;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
+import com.generalbytes.batm.server.extensions.extra.dash.sources.coinmarketcap.CMCData;
+import com.generalbytes.batm.server.extensions.extra.dash.sources.coinmarketcap.CMCMetaData;
 import java.math.BigDecimal;
-public class CMCMetaData {
-    public long timestamp;
-    public String error;
-}
-public class CMCQuote {
-    public BigDecimal price;
-    public BigDecimal volume_24h;
-    public BigDecimal market_cap;
-    public BigDecimal percent_change_1h;
-    public BigDecimal percent_change_24h;
-    public BigDecimal percent_change_7d;
-}
-public class CMCQuotes {
-    public CMCQuote USD;
-}
-public class CMCData {
-    private String id;
-    private String name;
-    private String symbol;
-    private String website_slug;
-    private BigDecimal rank;
-    private BigDecimal circulating_supply;
-    private BigDecimal total_supply;
-    private BigDecimal max_supply;
-    private CMCQuotes quotes;
-    private long last_updated;
+public class CMCTicker {
+    public CMCData jsondData;
+    public CMCMetaData metaData;
 
     public String getId() {
-        return id;
+        return jsondData.id;
     }
 
     public String getName() {
-        return name;
+        return jsondData.name;
     }
 
     public String getSymbol() {
-        return symbol;
+        return jsondData.symbol;
     }
 
     public String getWebsite_slug() {
-        return website_slug;
+        return jsondData.website_slug;
     }
 
     public BigDecimal getRank() {
-        return rank;
+        return jsondData.rank;
     }
 
     public BigDecimal getPrice() {
-        return quotes.USD.price;
+        return jsondData.quotes.USD.price;
     }
 
     public BigDecimal get_24h_volume() {
-         return quotes.USD.volume_24h;
+         return jsondData.quotes.USD.volume_24h;
     }
 
     public BigDecimal getMarket_cap() {
-         return quotes.USD.market_cap;
+         return jsondData.quotes.USD.market_cap;
     }
 
     public BigDecimal getCirculating_supply() {
-        return circulating_supply;
+        return jsondData.circulating_supply;
     }
 
     public BigDecimal getTotal_supply() {
-        return total_supply;
+        return jsondData.total_supply;
     }
 
     public BigDecimal getMax_supply() {
-        return max_supply;
+        return jsondData.max_supply;
     }
 
     public BigDecimal getPercent_change_1h() {
-        return quotes.USD.percent_change_1h;
+        return jsondData.quotes.USD.percent_change_1h;
     }
 
     public BigDecimal getPercent_change_24h() {
-        return quotes.USD.percent_change_24h;
+        return jsondData.quotes.USD.percent_change_24h;
     }
 
     public BigDecimal getPercent_change_7d() {
-        return quotes.USD.percent_change_7d;
+        return jsondData.quotes.USD.percent_change_7d;
     }
 
     public long getLast_updated() {
-        return last_updated;
+        return jsondData.last_updated;
     }
-}
-public class CMCTicker {
-    private CMCData jsondData;
-    private CMCMetaData metaData;
 }
